resolve debugging issues

+++ src/utils/cloundinary.js
@@ -1,23 +1,39 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import 'dotenv/config'
+
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
+
+// console.log(process.env.CLOUDINARY_CLOUD_NAME);
+// console.log(process.env.CLOUDINARY_API_KEY);
+// console.log(process.env.CLOUDINARY_API_SECRET);
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
+    console.log("Local file path:", localFilePath);
     if (!localFilePath) return null;
 
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
-    console.log("File uploaded succesfully. URL: ", response.url);
+    console.log("File uploaded successfully. URL: ", response.url);
+
+    // Check if the file exists before attempting to delete it
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
+    console.error("Error uploading to Cloudinary:", error);
+    
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     return null;
   }
 };
